Changed types

+++ BackEnd/src/main/java/com/stravacopy/backend/Model/SplitComparison.java
@@ -5,13 +5,13 @@ import lombok.Getter;
 @Getter
 public class SplitComparison {
     // Getters
-    private final int index;
-    private final int speedChange;
-    private final int heartRateChange;
-    private final int distanceChange;
+    private final Double index;
+    private final Double speedChange;
+    private final Double heartRateChange;
+    private final Double distanceChange;
     // more statistical comparisons...
 
-    public SplitComparison(int index, int speedChange, int heartRateChange, int distanceChange) {
+    public SplitComparison(Double index, Double speedChange, Double heartRateChange, Double distanceChange) {
         this.index = index;
         this.speedChange = speedChange;
         this.heartRateChange = heartRateChange;
